# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "API documentation for my Laravel application",
        "version": "1.0.0"
    },
    "paths": {
        "/api/actions/create": {
            "post": {
                "tags": [
                    "Actions"
                ],
                "summary": "Create a new action",
                "description": "Create a new action with camera_id, animal_type, and image_base64",
                "operationId": "5a0c0bb9a9a9a10ccec8f7519750f47d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "camera_id",
                                    "animal_type",
                                    "image_base64"
                                ],
                                "properties": {
                                    "camera_id": {
                                        "description": "ID of the camera",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "animal_type": {
                                        "description": "Animal type (0=Dog, 1=Pig, 2=Pigeon)",
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "image_base64": {
                                        "description": "Base64 encoded image",
                                        "type": "string",
                                        "example": "iVBORw0KGgoAAAANSUhEUgAABAAAA...=="
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Action created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Action successfully created"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "camera_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "animal_type": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "image_base64": {
                                                    "type": "string",
                                                    "example": "iVBORw0KGgoAAAANSUhEUgAABAAAA...=="
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-05-20T10:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-05-20T10:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, validation failed"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/actions/get/{id}": {
            "get": {
                "tags": [
                    "Actions"
                ],
                "summary": "Get Action by ID",
                "description": "Retrieve an action by its ID",
                "operationId": "3e9a9e037b9e96359b51c2a1e42e3ff1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Action retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "image_base64": {
                                            "type": "string",
                                            "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAlgAAAJYCAYAAAC+ZpjcAAAACXBIWXMAAA7EAAAOxAGVKw4bAAAgAElEQVR4nOy93a9uT3Ie9PSa7T1Hx0eTkZnMOMNoZE3MaBhbI9uEMODgHLEPwkIhQgIhghASd0iICyuKUCQQl1zlij8AIRAB8SGEfEP0czR27EzioBiifOGLKDH22B5b9njm+Pz27N8+q7l4u3o99XRVr/W++93mxi0dnXevj+pa1VVPV39UdcF1ygJgDX7PntNrS/s7+21/R+8cvZfxy89l7y2Ta9m70d96PeNP34nuR+WIfGe0Mj6P/K00Iz1g2rN3jshKaUV6uCer7J0Zj7P393jU8oeyynnUcq6sou/Zsw8Ef/N7l8oqup89z7xFtJWPP5TVH8oqov+Hsvr/WVZlwuDZ5e7u9VIaxVqBUoAPPvhaVPHR4gT05s3rThc4THvPWQEAvHnzehAk0d9zyPYUcXnz5rXStntH+Jx1Rk7uQp/5VN4OG4/KRuS+p4Cz6+tBuWeKP23/Ce0jdHeNh9u0VuBnfuZrw/clv7mEslLazZaidy6S1aRNI907S1aJ3Gd8271Dstrh3T27wzf/jQnvmZ6dLSujn7TpXmeU3V8ArzPE954N6rVDfDf6ER8XyYrlTraUdbhnyeru7jUO4uNFsjqgj0c7c+Z9hu2ZHJV+VLf7joD3iMbwXnBvuK76KLzP+rhdWV0B25W++53YEvPJJZXVTXBx1vkedk5KORlKQIMZij5W63MdGTtwwpM6B3Yv80hDBdvhV3/z/5Hh8TNaIiVY5Z79tr+1E2ZQWkoJaWXfAbnPshocwuD52Xdk9WlHZqC9AFgVSDDqibZ1CHwJ76pvM0cq0s1OR5yfpZTOO8tR22um53atDyK4lHLISdiV1RltepGsGu9..."
                                        },
                                        "animal_type": {
                                            "type": "string",
                                            "example": "Dog"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-05-20T10:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Action not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Actions",
            "description": "Actions"
        }
    ]
}